Change to a simpler command line syntax
To avoid conflict with hostnames we required the optional operation to be
specified as the last argument, which didn't feel quite right. Better to
remove explicit operations all together.

The get-operation were mostly redundant (since the set-operation did the
same thing if the hostname were already mapped).

The remove-operation is replaced with the new "d"-flag.